Swap component names
