+++ projects/web/src/pages/StandaloneApiReferencePage.vue
@@ -1,45 +1,62 @@
 <script setup lang="ts">
 import {
-  ApiReferenceBase,
+  ApiReference,
   type ReferenceConfiguration,
 } from '@scalar/api-reference'
+import { type ThemeId } from '@scalar/themes'
 import { reactive } from 'vue'
 
-import DevToolbar from '../components/DevToolbar.vue'
 import SlotPlaceholder from '../components/SlotPlaceholder.vue'
+
+// import preparsedContent from '../fixtures/specResult.json'
+
+const handleUpdateContent = (content: string) => {
+  window.localStorage.setItem('api-reference-content', content)
+}
 
 const configuration = reactive<ReferenceConfiguration>({
   theme: 'default',
   proxy: 'http://localhost:5051',
   isEditable: true,
   spec: {
-    url: 'https://raw.githubusercontent.com/outline/openapi/main/spec3.json',
+    content: window.localStorage.getItem('api-reference-content') ?? undefined,
+    // content: { openapi: '3.1.0', info: { title: 'Example' }, paths: {} },
+    // content: () => {
+    //   return { openapi: '3.1.0', info: { title: 'Example' }, paths: {} }
+    // },
+    // preparsedContent,
+    // url: 'https://raw.githubusercontent.com/outline/openapi/main/spec3.json',
+    // url: 'https://raw.githubusercontent.com/testimio/public-openapi/TES-14404-mobile-applications/api.yaml',
   },
   tabs: {
     initialContent: 'Swagger Editor',
   },
+  // hocuspocusConfiguration: {
+  //   name: 'document-1',
+  //   token: 'secret',
+  //   url: 'ws://localhost:1234',
+  // },
 })
 </script>
+
 <template>
-  <ApiReferenceBase :configuration="configuration">
-    <template #header>
-      <DevToolbar
-        :modelValue="configuration"
-        @update:modelValue="
-          (newConfiguration) => Object.assign(configuration, newConfiguration)
-        " />
+  <!-- <textarea
+    v-model="spec"
+    cols="30"
+    rows="10" /> -->
+  <!-- <select
+    @input="(event) => configuration.hocuspocusConfiguration ? configuration.hocuspocusConfiguration.name = (event.target as HTMLSelectElement).value : null"
+    @value="configuration.hocuspocusConfiguration?.name">
+    <option value="document-1">Document #1</option>
+    <option value="document-2">Document #2</option>
+    <option value="document-3">Document #3</option>
+  </select> -->
+  <ApiReference
+    :configuration="configuration"
+    @changeTheme="(theme: ThemeId) => (configuration.theme = theme)"
+    @updateContent="handleUpdateContent">
+    <template #footer>
+      <SlotPlaceholder>footer</SlotPlaceholder>
     </template>
-    <template #sidebar-start>
-      <SlotPlaceholder>sidebar-start</SlotPlaceholder>
-    </template>
-    <template #sidebar-end>
-      <SlotPlaceholder>sidebar-end</SlotPlaceholder>
-    </template>
-    <template #content-start>
-      <SlotPlaceholder>content-start</SlotPlaceholder>
-    </template>
-    <template #content-end>
-      <SlotPlaceholder>content-end</SlotPlaceholder>
-    </template>
-  </ApiReferenceBase>
+  </ApiReference>
 </template>
