+++ projects/web/src/pages/StartPage.vue
@@ -7,12 +7,14 @@ import StartPageLink from '../components/PageLink.vue'
     <div class="grid grid-cols-2 grid-rows-2 gap-4 p-6">
       <StartPageLink to="api-reference">
         <template #title>API Reference</template>
-        <template #description>Interactive API documentations</template>
+        <template #description>
+          Customizable Interactive API documentations
+        </template>
       </StartPageLink>
       <StartPageLink to="standalone-api-reference">
         <template #title>Standalone API Reference</template>
         <template #description>
-          Interactive API documentation without only basic configuration
+          Interactive API documentation with pre-configured search
         </template>
       </StartPageLink>
       <StartPageLink to="api-client">
